Version 2.2 Experimental update (may revert)
Decided to do away with copy.deepcopy function because it was really slowing down everything. Will test and revert back if it will not work properly on the test.
Please use this update with care as it may not work properly. This change is mostly in the MIDI.py core part of TMIDI so all functions may be affected.